fixes #5321: fix(visualization): Make results page load faster.
Because:
* The results page loads really slowly
* We're going to need the results available in the db for sorting/filtering

This commit:
* Caches the results in the db as a json blob w/ a daily task to update it
* Only embeds graphs when a collapsed section is expanded to speed up page load
