+++ app/experimenter/experiments/tests/test_changelog_utils/test_changelog_utils_nimbus.py
@@ -49,6 +49,7 @@ class TestNimbusExperimentChangeLogSerializer(TestCase):
                 "public_description": "",
                 "publish_status": NimbusExperiment.PublishStatus.IDLE.value,
                 "published_dto": None,
+                "results_data": None,
                 "reference_branch": None,
                 "risk_brand": None,
                 "risk_mitigation_link": "",
@@ -109,6 +110,7 @@ class TestNimbusExperimentChangeLogSerializer(TestCase):
                 "public_description": experiment.public_description,
                 "publish_status": experiment.publish_status,
                 "published_dto": None,
+                "results_data": None,
                 "risk_brand": experiment.risk_brand,
                 "risk_mitigation_link": experiment.risk_mitigation_link,
                 "risk_partner_related": experiment.risk_partner_related,
