+++ app/experimenter/jetstream/tests/test
@@ -1,0 +1,1 @@
+class TestCreateBugTask(MockRequestMixin, MockBugzillaMixin, TestCase):
